feat(knowledge,agent): 完成知识库前端管理、Agent知识库绑定及测试
- 修复 Agent 列表接口返回 tools/knowledges 为 null 的问题
- 修复知识库列表接口返回 documentCount 为 null 的问题
- 新增知识库文档管理独立页面，支持筛选、上传、编辑、删除、查看分块
- 新增向量召回测试页面，支持多知识库选择、Rerank 精排、置信度过滤
- 重构侧边栏导航，支持知识库管理子菜单展开/收起
- 优化 /api/v1/app/knowledge/search 接口，支持多知识库向量召回 + Rerank
- README.md 新增产品展示板块，补充 8 张菜单页截图

+++ ciff-knowledge/src/main/java/com/ciff/knowledge/facade/KnowledgeFacadeImpl.java
@@ -58,7 +58,7 @@ public class KnowledgeFacadeImpl implements KnowledgeFacade {
                     .flatMap(kid -> knowledgeChunkService.search(queryEmbedding, kid, topN * 2).stream())
                     .toList();
         } else {
-            candidates = knowledgeChunkService.search(queryEmbedding, null, topN * 3);
+            candidates = knowledgeChunkService.search(queryEmbedding, (Long) null, topN * 3);
         }
 
         if (candidates.isEmpty()) {
